Updated 1.01
Shopping Cart Buttons Updated - Some of the shopping cart buttons for the wtw-shopping plugin were not properly textured, making it difficult to see the lettering. We made the casing blue, button white, and text black. This gives a high contrast in the lettering and background. This update includes the Add to Cart, Read More, Buy Now, View Cart, and Check Out buttons.