Updated with WalkTheWeb 3.4.4
Admin Store 3D Sign Colors - Product Displays for 3D Shopping Store Products basically display the image, title, and price with links to read more and add to cart. There are times when you want to use a 3D Model to show the product instead of the product image, but you still want the title, price, read more, and add to cart. The Display with No Image is great for these 3D Models. It is like a stage for the model that matches the rest of your Product Displays.

Retired the retro box Product Displays - Since we now have 3 clean product Displays to build from, we decided to retire the old retro box displays. The decision was made because they used different code and were built with less quality. The new Product Displays are easy to read, resize, and color. 

Fixed Store Categories Display and Hovers - Set a minimum width so that the store Categories Hovers have room to work. In addition, fixed the color settings by making sure it selects the image frame of the object instead of the placement box (which should remain invisible).

Keyboard Input Additions and Changes - In order to type into a mesh on the screen, we needed to identify the mesh and pause the keyboard input for avatar movement. We use the global variable WTW.selectedMoldName to set the id of the active mesh, then altered the keyDown function in the wtw_inputs.js file to perform the keyPressedAdd function for movement when no mesh is active (selected). In addition, we added a function to process the Key Down from within plugins. Finally, we moved the preventDefault to they scrollbox activity. This can be manually added if needed on the plugin key down activity.

Cleared Selected Mold Name - When you click on a mesh to make it active (like the 3D Store Product Search Textbox), it sets the global variable WTW.selectedMoldName to the id of the mold and stops processing the Avatar Movement. There are now multiple ways to clear the active mold and return to movement; Press ESC or you can click or touch anywhere else on the canvas 3D Scene Window. We added code to set WTW.selectedMoldName=' in a number of touch, mouse click, and key down functions in the wtw_input.js page.

Created Clear Selected Mold Function - Seems that there was more to do than just clear the global variable WTW.selectedMoldName. Therefore, I created a function called clearSelectedMold and allow plugins to tap into this function to be executed before the variable is cleared. For 3D Shopping we have a text box you can type into when you select it. If it is active, it shows a blinking cursor. We added code to clear the blinker timer and take away the cursor in the text box when you close the selected mold.

Added 3D Store Product Keyword Search - Previously, the only way to look through products were to select a category. Now we added a Keyword Product Search! Just click the blank to see a cursor, type, and click the Search button. The product displays that are set to allow search to override (default setting) will update to the search results. If no results are returned, the product displays stay on the previous items.

More Store Product Search Kiosk Choices - We added a total of four search kiosk configurations. Just add a Product Search from under Add 3D Store Objects and select which kiosk you would like from the dropdown list. All Kiosks can be set to any color highlighting to match any 3D Store. You can even have more than one kiosk in a 3D Store, they will all return the searched items if they are found.

Shadow Performance Improvement - We implemented the Use Percentage Closer Filtering on shadows to help improve performance. We will be experimenting with a number of other shadow related settings to balance quality and performance.