Updated with WalkTheWeb 3.4.9
Added Language Translation to 3D Plugins - Added the ability to add a translation file to a languages folder under the plugin root t translate any labels and terms used in the plugin. We also added the translation PHP function __() to the wtwplugins class. 

Alternative Language Spanish - The focus of the language translation is the menu system and the first alternative language included is Spanish. It starts with the json language file in the /core/languages folder and has the addition of a language file(s) in the plugins languages subfolders. This stage of release did not fully translate all of the menu forms or full page forms.